fix(apiserver-duration): compute average of duration
without the fix, the dashboards show the cumulative duration
of all buckets, which doesn't make much sense.
with the fix, we divide the total duration by the number of requests,
which yields the average latency. note that we could also have used
`_duration_seconds_bucket{le="+Inf"}` instead of `duration_seconds_sum`

Signed-off-by: Clément Nussbaumer <clement.nussbaumer@postfinance.ch>